Remove the Solution_Old from my_Clone routine.
Since only the Solution at current and at previout time are needed to
carry out a time integration step, the Solution_Old variable does not
have to be copies. Solution_Old is only used for iterating a Solution at
the current time step, it contains the old iterate and gets overwritten
in every iteration.