BUGFIX for my_Access. Run braid wrapper tests.
The main routine runs the braid wrapper test function for my_Access.
The routine my_Access now passes the current time step number to SU2 so
that the output files carry the correct names. Furthermore, the output
routine takes the nZone variable as the last argument (which is 1 in
this case), not the ZONE_0 value!!