Add Verbosity option to the Testcase config file.
The laminar airfoil shedding config file now has the option for
verbosity of the solver. It is set to NONE. The output routines of SU2
are changed such that they do only print messages on screen, if the
verbosity option is medium of high. Only the MeanFlowIteration is
changed so far.