Run xBraid in time serial (pt=1, max_level=1).
The SU2 code is called from XBraid's phi function. xBraid runs in time
serial mode if pt=1 and max_level=1 is chosen.
For px=4, pt=1, i.e. all 4 processors are used for spatial
parallelization, the code produced the same result with the pure SU2
dual-time stepping code for 4 processors.
However, using 8 cores for SU2 with px=8, pt=1 and comparing to the
4-processor dual-time stepping reference, the relative errors are around
1e-4.
Don't know, if that is ok.