BUGFIX in my_Phi function. Set the iExtIter.
The iExtIter variable has to be set correctly inside the my_Phi function
before calling SU2's driver->run, because if iExtIter is zero, the
solution_time_n1 and solution_time_n get overwritten to freestream.

This is probably the reason for the not-matching result of the serial
run of XBraid on multiple cores. TODO: Check this!