Add structure TwoStepSolution with con-/destructor.
The braid_Vector now contains a pointer to a TwoStepSolution structure
which contains the solution lists at time n and n-1. Constructor and
Destructor are added for the TwoStepSolution struct which allocate and
deallocate the memory for the solution lists.

The braid_Vector does not have a constructor and destructor, because the
primal solution lists are allocated and freed with the corresponding
my_Init and my_Free functions from xBraid.

However, the destructor of the TwoStepSolution is needed for the
implementation of the adjoint tape. This will contain shared pointers to
adjoint TwoStepSolutions 'Solution_b'. During primal computations,
shared pointers will be copies and pushed onto the adjointtape, in the
reverse run, these pointers will be poped out for adjoint computation.
If no pointer is left, the shared_ptr will call default destructor of
the TwoStepSolution, i.e. will free the memory for the adjoint.