Testing black box sensitivity of one surface point.
Finite differencing for one surface point: Perturb a coordinate of the
surface inside the my_Step_adjoint routine. Check the gradient of ONE
step of SU2->Run:
The relative error of the AD blackbox gradient with finite difference is
around 3% (eps=1r-5).

TODO: Check the derivative for the whole my_Step_adjoint routine.